requirements.txt
streamlit>=1.32.0
google-generativeai>=0.8.3
python-dotenv>=1.0.0
langchain>=0.1.0
python-docx>=1.1.0
reportlab>=4.2.0
